# frontend/src/generated/openapi.json
{
  "components": {
    "schemas": {
...
      "Artifact": {
        "description": "Artifact file metadata.",
        "properties": {
          "exists": {
            "description": "Whether the file exists",
            "title": "Exists",
            "type": "boolean"
          },
          "name": {
            "description": "File name",
            "title": "Name",
            "type": "string"
          },
          "path": {
            "description": "File path",
            "title": "Path",
            "type": "string"
          },
          "size": {
            "description": "Size in bytes",
            "title": "Size",
            "type": "integer"
          },
          "type": {
            "description": "File type: csv, json, txt, etc.",
            "title": "Type",
            "type": "string"
          }
        },
        "required": [
          "name",
          "path",
          "type",
          "size",
          "exists"
        ],
        "title": "Artifact",
        "type": "object"
      },
      "BacktestMetrics": {
        "additionalProperties": true,
        "description": "Backtest summary metrics.",
        "properties": {
          "annual_return": {
            "description": "Annualized return",
            "title": "Annual Return",
            "type": "number"
          },
          "final_value": {
            "description": "Ending portfolio value",
            "title": "Final Value",
            "type": "number"
          },
          "max_drawdown": {
            "description": "Max drawdown",
            "title": "Max Drawdown",
            "type": "number"
          },
          "sharpe": {
            "description": "Sharpe ratio",
            "title": "Sharpe",
            "type": "number"
          },
          "total_return": {
            "description": "Total return",
            "title": "Total Return",
            "type": "number"
          },
          "trade_count": {
            "description": "Number of trades",
            "title": "Trade Count",
            "type": "integer"
          },
          "win_rate": {
            "description": "Win rate",
            "title": "Win Rate",
            "type": "number"
          }
        },
        "required": [
          "final_value",
          "total_return",
          "annual_return",
          "max_drawdown",
          "sharpe",
          "win_rate",
          "trade_count"
        ],
        "title": "BacktestMetrics",
        "type": "object"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "title": "Detail",
            "type": "array"
          }
        },
        "title": "HTTPValidationError",
        "type": "object"
      },
...
      "RAGSelection": {
        "description": "RAG routing result.",
        "properties": {
          "selected_api": {
            "description": "Selected API code",
            "title": "Selected Api",
            "type": "string"
          },
          "selected_name": {
            "description": "Selected API name",
            "title": "Selected Name",
            "type": "string"
          },
          "selected_score": {
            "description": "Match score",
            "title": "Selected Score",
            "type": "number"
          }
        },
        "required": [
          "selected_api",
          "selected_name",
          "selected_score"
        ],
        "title": "RAGSelection",
        "type": "object"
      },
...
      "RunInfo": {
        "description": "Compact run row for list views.",
        "properties": {
          "codes": {
            "items": {
              "type": "string"
            },
            "title": "Codes",
            "type": "array"
          },
          "created_at": {
            "title": "Created At",
            "type": "string"
          },
          "end_date": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "End Date"
          },
          "prompt": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Prompt"
          },
          "run_id": {
            "title": "Run Id",
            "type": "string"
          },
          "sharpe": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Sharpe"
          },
          "start_date": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Start Date"
          },
          "status": {
            "title": "Status",
            "type": "string"
          },
          "total_return": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Total Return"
          }
        },
        "required": [
          "run_id",
          "status",
          "created_at"
        ],
        "title": "RunInfo",
        "type": "object"
      },
      "RunResponse": {
        "description": "API response payload for a single run.",
        "properties": {
          "artifacts": {
            "description": "Run artifacts",
            "items": {
              "$ref": "#/components/schemas/Artifact"
            },
            "title": "Artifacts",
            "type": "array"
          },
          "artifacts_equity_csv": {
            "anyOf": [
              {
                "items": {
                  "additionalProperties": true,
                  "type": "object"
                },
                "type": "array"
              },
              {
                "type": "null"
              }
            ],
            "description": "Full equity rows",
            "title": "Artifacts Equity Csv"
          },
          "artifacts_metrics_csv": {
            "anyOf": [
              {
                "items": {
                  "additionalProperties": true,
                  "type": "object"
                },
                "type": "array"
              },
              {
                "type": "null"
              }
            ],
            "description": "Full metrics rows",
            "title": "Artifacts Metrics Csv"
          },
          "artifacts_trades_csv": {
            "anyOf": [
              {
                "items": {
                  "additionalProperties": true,
                  "type": "object"
                },
                "type": "array"
              },
              {
                "type": "null"
              }
            ],
            "description": "Full trade rows",
            "title": "Artifacts Trades Csv"
          },
          "elapsed_seconds": {
            "description": "Execution time in seconds",
            "title": "Elapsed Seconds",
            "type": "number"
          },
          "equity_curve": {
            "anyOf": [
              {
                "items": {
                  "additionalProperties": true,
                  "type": "object"
                },
                "type": "array"
              },
              {
                "type": "null"
              }
            ],
            "description": "Equity preview",
            "title": "Equity Curve"
          },
          "indicator_series": {
            "anyOf": [
              {
                "additionalProperties": {
                  "additionalProperties": {
                    "items": {
                      "additionalProperties": true,
                      "type": "object"
                    },
                    "type": "array"
                  },
                  "type": "object"
                },
                "type": "object"
              },
              {
                "type": "null"
              }
            ],
            "description": "Grouped indicator overlays",
            "title": "Indicator Series"
          },
          "llm_usage": {
            "anyOf": [
              {
                "additionalProperties": true,
                "type": "object"
              },
              {
                "type": "null"
              }
            ],
            "description": "Provider-reported AgentLoop usage summary",
            "title": "Llm Usage"
          },
          "metrics": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/BacktestMetrics"
              },
              {
                "type": "null"
              }
            ],
            "description": "Backtest metrics"
          },
          "planner_output": {
            "anyOf": [
              {
                "additionalProperties": true,
                "type": "object"
              },
              {
                "type": "null"
              }
            ],
            "description": "Planner output",
            "title": "Planner Output"
          },
          "price_series": {
            "anyOf": [
              {
                "additionalProperties": {
                  "items": {
                    "additionalProperties": true,
                    "type": "object"
                  },
                  "type": "array"
                },
                "type": "object"
              },
              {
                "type": "null"
              }
            ],
            "description": "Grouped OHLC series",
            "title": "Price Series"
          },
          "rag_selection": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/RAGSelection"
              },
              {
                "type": "null"
              }
            ],
            "description": "Selected RAG metadata"
          },
          "reason": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "description": "Failure reason when available",
            "title": "Reason"
          },
          "run_card": {
            "anyOf": [
              {
                "additionalProperties": true,
                "type": "object"
              },
              {
                "type": "null"
              }
            ],
            "description": "Trust Layer run card payload",
            "title": "Run Card"
          },
          "run_context": {
            "anyOf": [
              {
                "additionalProperties": true,
                "type": "object"
              },
              {
                "type": "null"
              }
            ],
            "description": "Normalized request context",
            "title": "Run Context"
          },
          "run_directory": {
            "description": "Run directory path",
            "title": "Run Directory",
            "type": "string"
          },
          "run_id": {
            "description": "Run identifier",
            "title": "Run Id",
            "type": "string"
          },
          "run_logs": {
            "anyOf": [
              {
                "items": {
                  "additionalProperties": true,
                  "type": "object"
                },
                "type": "array"
              },
              {
                "type": "null"
              }
            ],
            "description": "Structured stdout/stderr lines",
            "title": "Run Logs"
          },
          "run_stage": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "description": "UI-facing run stage",
            "title": "Run Stage"
          },
          "status": {
            "description": "Run status: success, failed, aborted",
            "title": "Status",
            "type": "string"
          },
          "strategy_spec": {
            "anyOf": [
              {
                "additionalProperties": true,
                "type": "object"
              },
              {
                "type": "null"
              }
            ],
            "description": "Strategy specification",
            "title": "Strategy Spec"
          },
          "trade_log": {
            "anyOf": [
              {
                "items": {
                  "additionalProperties": true,
                  "type": "object"
                },
                "type": "array"
              },
              {
                "type": "null"
              }
            ],
            "description": "Trade preview",
            "title": "Trade Log"
          },
          "trade_markers": {
            "anyOf": [
              {
                "items": {
                  "additionalProperties": true,
                  "type": "object"
                },
                "type": "array"
              },
              {
                "type": "null"
              }
            ],
            "description": "Trade markers for charts",
            "title": "Trade Markers"
          },
          "validation": {
            "anyOf": [
              {
                "additionalProperties": true,
                "type": "object"
              },
              {
                "type": "null"
              }
            ],
            "description": "Statistical validation results",
            "title": "Validation"
          }
        },
        "required": [
          "status",
          "run_id",
          "elapsed_seconds",
          "run_directory"
        ],
        "title": "RunResponse",
        "type": "object"
      },
...
      "ValidationError": {
        "properties": {
          "ctx": {
            "title": "Context",
            "type": "object"
          },
          "input": {
            "title": "Input"
          },
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "title": "Location",
            "type": "array"
          },
          "msg": {
            "title": "Message",
            "type": "string"
          },
          "type": {
            "title": "Error Type",
            "type": "string"
          }
        },
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError",
        "type": "object"
      },
...
    },
    "securitySchemes": {
      "HTTPBearer": {
        "scheme": "bearer",
        "type": "http"
      }
    }
  },
  "info": {
    "description": "Vibe-Trading API: natural-language finance research, backtesting, and swarm workflows",
    "title": "Vibe-Trading API",
    "version": "0.1.10"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/api/v1/runs": {
      "get": {
        "operationId": "v1_list_runs_api_v1_runs_get",
        "parameters": [
          {
            "in": "query",
            "name": "limit",
            "required": false,
            "schema": {
              "default": 20,
              "title": "Limit",
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/RunInfo"
                  },
                  "title": "Response V1 List Runs Api V1 Runs Get",
                  "type": "array"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "V1 List Runs",
        "tags": [
          "v1"
        ]
      }
    },
    "/api/v1/runs/{run_id}": {
      "get": {
        "operationId": "v1_get_run_result_api_v1_runs__run_id__get",
        "parameters": [
          {
            "in": "path",
            "name": "run_id",
            "required": true,
            "schema": {
              "title": "Run Id",
              "type": "string"
            }
          },
          {
            "description": "Opt in to chart payloads for a single symbol",
            "in": "query",
            "name": "chart_symbol",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "description": "Opt in to chart payloads for a single symbol",
              "title": "Chart Symbol"
            }
          },
          {
            "description": "Optional chart payload mode. Use 'summary' to omit chart rows and trade markers.",
            "in": "query",
            "name": "chart_payload",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "description": "Optional chart payload mode. Use 'summary' to omit chart rows and trade markers.",
              "title": "Chart Payload"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RunResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "V1 Get Run Result",
        "tags": [
          "v1"
        ]
      }
    },
    "/api/v1/runs/{run_id}/code": {
      "get": {
        "operationId": "v1_get_run_code_api_v1_runs__run_id__code_get",
        "parameters": [
          {
            "in": "path",
            "name": "run_id",
            "required": true,
            "schema": {
              "title": "Run Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {}
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "V1 Get Run Code",
        "tags": [
          "v1"
        ]
      }
    },
    "/api/v1/runs/{run_id}/pine": {
      "get": {
        "operationId": "v1_get_run_pine_api_v1_runs__run_id__pine_get",
        "parameters": [
          {
            "in": "path",
            "name": "run_id",
            "required": true,
            "schema": {
              "title": "Run Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {}
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "V1 Get Run Pine",
        "tags": [
          "v1"
        ]
      }
    },
...
    "/runs": {
      "get": {
        "operationId": "list_runs_runs_get",
        "parameters": [
          {
            "in": "query",
            "name": "limit",
            "required": false,
            "schema": {
              "default": 20,
              "title": "Limit",
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/RunInfo"
                  },
                  "title": "Response List Runs Runs Get",
                  "type": "array"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "List Runs"
      }
    },
    "/runs/{run_id}": {
      "get": {
        "operationId": "get_run_result_runs__run_id__get",
        "parameters": [
          {
            "in": "path",
            "name": "run_id",
            "required": true,
            "schema": {
              "title": "Run Id",
              "type": "string"
            }
          },
          {
            "description": "Opt in to chart payloads for a single symbol",
            "in": "query",
            "name": "chart_symbol",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "description": "Opt in to chart payloads for a single symbol",
              "title": "Chart Symbol"
            }
          },
          {
            "description": "Optional chart payload mode. Use 'summary' to omit chart rows and trade markers.",
            "in": "query",
            "name": "chart_payload",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "description": "Optional chart payload mode. Use 'summary' to omit chart rows and trade markers.",
              "title": "Chart Payload"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RunResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Get Run Result"
      }
    },
    "/runs/{run_id}/code": {
      "get": {
        "operationId": "get_run_code_runs__run_id__code_get",
        "parameters": [
          {
            "in": "path",
            "name": "run_id",
            "required": true,
            "schema": {
              "title": "Run Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {}
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Get Run Code"
      }
    },
    "/runs/{run_id}/pine": {
      "get": {
        "operationId": "get_run_pine_runs__run_id__pine_get",
        "parameters": [
          {
            "in": "path",
            "name": "run_id",
            "required": true,
            "schema": {
              "title": "Run Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {}
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Get Run Pine"
      }
    },
...
  }
}